added authentication and authorization

+++ db/queries/requests/delete_task_query.sql
@@ -1,4 +1,5 @@
 -- Deletes a task
 DELETE FROM tasks
 WHERE
-    tasks.id = ?;
+    tasks.id = ?
+    AND tasks.userid = ?;
